BitKeeper/deleted/.del-xarQuery.php~704b38a9: 	#  Delete: xarincludes/xarQuery.php admin/accounting.php: 	#  more changes... admin/banner_manager.php: 	#  more changes... admin/banner_statistics.php: 	#  more changes... admin/categories_view.php: 	#  more changes... admin/configuration.php: 	#  more changes... admin/content_preview.php: 	#  more changes... admin/create_account.php: 	#  more changes... admin/includes/functions/html_graphs.php: 	#  more changes... admin/includes/functions/html_output.php: 	#  more changes... admin/includes/functions/sessions.php: 	#  more changes... admin/includes/graphs/banner_daily.php: 	#  more changes... admin/includes/graphs/banner_infobox.php: 	#  more changes... admin/includes/graphs/banner_monthly.php: 	#  more changes... admin/includes/graphs/banner_yearly.php: 	#  more changes... admin/includes/modules/customer_memo.php: 	#  more changes... admin/includes/modules/group_prices.php: 	#  more changes... admin/includes/modules/newsletters/newsletter.php: 	#  more changes... admin/includes/modules/newsletters/product_notification.php: 	#  more changes... admin/includes/modules/security_check.php: 	#  more changes... admin/index.php: 	#  more changes... admin/mail.php: 	#  more changes... admin/new_categorie.php: 	#  more changes... admin/new_product.php: 	#  more changes... admin/newsletters.php: 	#  more changes... admin/orders_status.php: 	#  more changes... admin/popup_image.php: 	#  more changes... admin/print_order.php: 	#  more changes... admin/print_packingslip.php: 	#  more changes... admin/stats_stock_warning.php: 	#  more changes... checkout_process.php: 	#  more changes... download.php: 	#  more changes... includes/boxes.php: 	#  more changes... includes/classes/banktransfer_validation.php: 	#  more changes... includes/classes/currencies.php: 	#  more changes... includes/classes/language.php: 	#  more changes... includes/classes/order.php: 	#  more changes... includes/classes/shopping_cart.php: 	#  more changes... includes/classes/split_page_results.php: 	#  more changes... includes/column_right.php: 	#  more changes... includes/counter.php: 	#  more changes... includes/functions/sessions.php: 	#  more changes... includes/general.php: 	#  more changes... includes/modules/address_book_details.php: 	#  more changes... includes/modules/also_purchased_products.php: 	#  more changes... includes/modules/checkout_new_address.php: 	#  more changes... includes/modules/downloads.php: 	#  more changes... includes/modules/graduated_prices.php: 	#  more changes... includes/modules/metatags.php: 	#  more changes... includes/modules/new_products.php: 	#  more changes... includes/modules/product_attributes.php: 	#  more changes... includes/modules/product_info.php: 	#  more changes... includes/modules/product_listing.php: 	#  more changes... includes/modules/product_reviews.php: 	#  more changes... includes/modules/products_media.php: 	#  more changes... includes/modules/upcoming_products.php: 	#  more changes... includes/write_customers_status.php: 	#  more changes... includes/xapplication_top.php: 	#  more changes... index.php: 	#  more changes... media_content.php: 	#  more changes... popup_image.php: 	#  more changes... print_order.php: 	#  more changes... print_product_info.php: 	#  more changes... redirect.php: 	#  more changes... reviews.php: 	#  more changes... send_order.php: 	#  more changes... xaradmin/backup.php: 	#  more changes... xaradmin/categories.php: 	#  more changes... xaradmin/categories_screen.php: 	#  more changes... xaradmin/categoriesold.php: 	#  more changes... xaradmin/content_manager.php: 	#  more changes... xaradmin/countries.php: 	#  more changes... xaradmin/countriesold.php: 	#  more changes... xaradmin/create_account.php: 	#  more changes... xaradmin/create_accountold.php: 	#  more changes... xaradmin/currencies.php: 	#  more changes... xaradmin/currenciesold.php: 	#  more changes... xaradmin/customers.php: 	#  more changes... xaradmin/customers_status.php: 	#  more changes... xaradmin/customers_statusold.php: 	#  more changes... xaradmin/customersold.php: 	#  more changes... xaradmin/edit_account.php: 	#  more changes... xaradmin/geo_zones.php: 	#  more changes... xaradmin/geo_zonesold.php: 	#  more changes... xaradmin/languages.php: 	#  more changes... xaradmin/languagesold.php: 	#  more changes... xaradmin/manufacturers.php: 	#  more changes... xaradmin/manufacturersold.php: 	#  more changes... xaradmin/module_newsletter.php: 	#  more changes... xaradmin/modules.php: 	#  more changes... xaradmin/new_attributes.php: 	#  more changes... xaradmin/orders.php: 	#  more changes... xaradmin/product_screen.php: 	#  more changes... xaradmin/products_attributes.php: 	#  more changes... xaradmin/products_attributesold.php: 	#  more changes... xaradmin/products_expected.php: 	#  more changes... xaradmin/reviews.php: 	#  more changes... xaradmin/specials.php: 	#  more changes... xaradmin/stats_customers.php: 	#  more changes... xaradmin/stats_products_purchased.php: 	#  more changes... xaradmin/stats_products_viewed.php: 	#  more changes... xaradmin/tax_classes.php: 	#  more changes... xaradmin/tax_classesold.php: 	#  more changes... xaradmin/tax_rates.php: 	#  more changes... xaradmin/tax_ratesold.php: 	#  more changes... xaradmin/whos_online.php: 	#  more changes... xaradmin/zones.php: 	#  more changes... xaradminapi/get_status_users.php: 	#  more changes... xaradminapi/get_user_info.php: 	#  more changes... xaradminapi/languages.php: 	#  more changes... xaradminapi/load_configuration.php: 	#  more changes... xaradminapi/remove_product.php: 	#  more changes... xarblocks/admin.php: 	#  more changes... xarblocks/best_sellers.php: 	#  more changes... xarblocks/categories.php: 	#  more changes... xarblocks/content.php: 	#  more changes... xarblocks/currencies.php: 	#  more changes... xarblocks/information.php: 	#  more changes... xarblocks/manufacturer_info.php: 	#  more changes... xarblocks/manufacturers.php: 	#  more changes... xarinit.php: 	#  more changes...

+++ includes/xapplication_top.php
@@ -95,7 +95,7 @@
   // set the application parameters
   $configuration_query = new xenQuery('select configuration_key as cfgKey, configuration_value as cfgValue from ' . TABLE_CONFIGURATION);
       $q = new xenQuery();
-      $q->run();
+      if(!$q->run()) return;
   while ($configuration = $q->output()) {
     define($configuration['cfgKey'], $configuration['cfgValue']);
   }
@@ -338,7 +338,7 @@
                                 xarRedirectResponse(xarModURL('commerce','user',(FILENAME_ADVANCED_SEARCH_RESULT, 'keywords=' . $_POST['quickie'], 'NONSSL'));
                               }
       $q = new xenQuery();
-      $q->run();
+      if(!$q->run()) return;
                               $quickie = $q->output();
                               if (xtc_has_product_attributes($quickie['products_id'])) {
                                 xarRedirectResponse(xarModURL('commerce','user','product_info', 'products_id=' . $quickie['products_id'], 'NONSSL'));
@@ -374,7 +374,7 @@
           for ($i = 0, $n = sizeof($notify); $i < $n; $i++) {
             $check_query = new xenQuery("select count(*) as count from " . TABLE_PRODUCTS_NOTIFICATIONS . " where products_id = '" . $notify[$i] . "' and customers_id = '" . $_SESSION['customer_id'] . "'");
       $q = new xenQuery();
-      $q->run();
+      if(!$q->run()) return;
             $check = $q->output();
             if ($check['count'] < 1) {
               new xenQuery("insert into " . TABLE_PRODUCTS_NOTIFICATIONS . " (products_id, customers_id, date_added) values ('" . $notify[$i] . "', '" . $_SESSION['customer_id'] . "', now())");
@@ -390,7 +390,7 @@
         if (isset($_SESSION['customer_id']) && isset($_GET['products_id'])) {
           $check_query = new xenQuery("select count(*) as count from " . TABLE_PRODUCTS_NOTIFICATIONS . " where products_id = '" . $_GET['products_id'] . "' and customers_id = '" . $_SESSION['customer_id'] . "'");
       $q = new xenQuery();
-      $q->run();
+      if(!$q->run()) return;
           $check = $q->output();
           if ($check['count'] > 0) {
             new xenQuery("delete from " . TABLE_PRODUCTS_NOTIFICATIONS . " where products_id = '" . $_GET['products_id'] . "' and customers_id = '" . $_SESSION['customer_id'] . "'");
